Move RTV utils into subdirectory
